% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{aquastat_rounder}
\alias{aquastat_rounder}
\title{Perform UN-style rounding}
\usage{
aquastat_rounder(x)
}
\arguments{
\item{x}{number to round}
}
\value{
x, but rounded
}
\description{
Rounds numbers greater than 1000 to no decimals,
greater than 100 to one decimal, etc.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
