% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{button_maker}
\alias{button_maker}
\title{Button maker}
\usage{
button_maker(value = NULL, subtitle = NULL, size = "md", icon = NULL,
  type = "warning", link = NULL, units = NULL, hover = NULL)
}
\description{
Button maker
}
