% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilemaker.R
\name{ButtonMaker}
\alias{ButtonMaker}
\title{ButtonMaker}
\usage{
ButtonMaker(Color = 1, Size = 4, Value, Subtitle = "", Link = "",
  Icon = "", Units = "", Target = 0, ThresholdHigh = 0,
  ThresholdLow = 0, Hover = "", alpha = 0.5, Former = Value)
}
\arguments{
\item{Color}{Optional numeric 1-6, corresponding to the colors specified in the bootstrap css classes:
\"success\",  \"warning\", \"danger\", \"info\", \"primary\", \"default\"}

\item{Size}{Optional numeric 1-4, corresponding to the sizes specified in the bootstrap css classes:
\"xs\",\"sm\",\"md\",\"lg\")}

\item{Value}{The numeric value you want to highlight (the main enchilada)}

\item{Subtitle}{Optional subtext that should appear under the value}

\item{Link}{Optional hyperlink that should be followed on click}

\item{Icon}{Optional glyphicon that should be displayed from http://getbootstrap.com/components/}

\item{Units}{Optional units that should be displayed after Value}

\item{Target}{Optional target that the value should be compared against. Use with ThresholdHigh and THresholdLow}

\item{ThresholdHigh}{Optional border between \"green\" and \"orange\". Use w/ Target and ThresholdLow. This value represents the RATIO
of the VALUE to the TARGET that, if above the ThresholdHigh will show as green, and if not, as orange}

\item{ThresholdLow}{Optional border between \"orange\" and \"red\". Use w/ Target and ThresholdLow. This value represents the RATIO
of the VALUE to the TARGET that, if above the ThresholdHigh will show as orange, and if not, as red}

\item{Hover}{Optional tooltip, or text that will show up when a user rests their mouse over the button.}

\item{alpha}{Optional transparency coefficient for the icon, a decimal from 0-1.}

\item{Former}{Optional value to compare against the current value. Will show up as a little arrow pointig the percent change}
}
\value{
Returns a character string containing HTML code to show the button, assuming the appropriate CSS elements will be available downstream
}
\description{
This function crafts the actual button per se, including the specific
aesthetic traits for each button
}
\examples{
# ADD EXAMPLES HERE
# Button1 <- ButtonMaker(Color = 2,Value = 3.57,Subtitle = "Times apple eaten")
# Button2 <- ButtonMaker(Color = 3,Value = 13.7,Subtitle = "Nutritional value")
# Button3 <- ButtonMaker(Color = 4,Value = 1,Subtitle = "Yumminess factor")
# TileMaker::ButtonMaker(Color = 5,Size=1,Value = 5,Subtitle = "Inconsistencies")
# Button1;Button2;Button3;Button4

}
