% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{div_maker}
\alias{div_maker}
\title{Div maker}
\usage{
div_maker(subtitle = NULL, textModifier = "h2", ...)
}
\arguments{
\item{subtitle}{The text heading of this row of buttons}

\item{textModifier}{Optional css category of "large" text. In this case, subtitle. Default=h2}

\item{...}{\code{solo_box OR solo_gradient_box} elements.}
}
\description{
This function takes strings created with the function `ButtonMaker` and makes an HTML div suitable for inclusion in other HTML code,
or for inclusion within the function of this package `TileMaker`.
}
\examples{
div_maker(title = "Quantativity factors",
          solo_gradient_box(value = 70),
          solo_box(value=34))
}
