% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilemaker.R
\name{DivMaker}
\alias{DivMaker}
\title{DivMaker}
\usage{
DivMaker(Title = "", Buttons)
}
\arguments{
\item{Title}{The title for this row of buttons}

\item{Buttons}{The Buttons that you want inserted into this row. If you have more than one button, use paste(Button1,Button2)}
}
\value{
Returns an HTML string containing \"div\" elements. Beware of using these in Shiny... a it might break the container.
}
\description{
This function takes strings created with the function `ButtonMaker` and makes an HTML div suitable for inclusion in other HTML code,
or for inclusion within the function of this package `TileMaker`.
}
\examples{
# ADD EXAMPLES HERE
Button1 <- ButtonMaker(Color = 2, Value = 3.57, Subtitle = "B")
Button2 <- ButtonMaker(Color = 3, Value = 13.7, Subtitle = "Nutritional value")
DivMaker(Title = "Quantativity factors", Buttons = paste(Button1, Button2))
}
