% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{file_maker}
\alias{file_maker}
\title{file maker}
\usage{
file_maker(title = NULL, ...,
  css = "https://bootswatch.com/4/flatly/bootstrap.css", file = NULL)
}
\arguments{
\item{title}{Title.}

\item{...}{\code{div_maker} elements.}

\item{css}{A string indicating css url, for final installations pls save the css file locally.}

\item{file}{Optional filename if you desire to save the file. Should end with ".html"}
}
\description{
Function 3 of 3, the last step. This function grabs the Divs created by `DivMaker` and
combines them into an html file.
}
