% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{finisher}
\alias{finisher}
\title{file maker}
\usage{
finisher(title = NULL, ...,
  css = "https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css",
  file = NULL, textModifier = "h1")
}
\arguments{
\item{title}{Title.}

\item{...}{\code{div_maker} elements.}

\item{css}{A string indicating css url, for final installations pls save the css file locally.}

\item{file}{Optional filename if you desire to save the file. By default we are using the 3.3.7 bootstrap CDN because
they support icons, but some others that might be interesting to you are:
https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css or https://bootswatch.com/4/flatly/bootstrap.css}

\item{textModifier}{Optional css category of "large" text. In this case, title. Default=h1}
}
\description{
Function 3 of 3, the last step. This function grabs the Divs created by `DivMaker` and
combines them into an html file.
}
