% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{ico}
\alias{ico}
\title{ico}
\usage{
ico(x, chevron = FALSE)
}
\arguments{
\item{x}{Icon name. See https://getbootstrap.com/docs/3.3/components/}

\item{chevron}{binary to denote whether there is a former value to compare against or not.}
}
\description{
Auxiliary function to generate icons
}
