% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{prettify}
\alias{prettify}
\title{apply pretty format or not}
\usage{
prettify(x, pretty = NULL)
}
\arguments{
\item{x}{thingie to evaluate (whether it should be prettified or not)}

\item{pretty}{inheriting the value that should be used as a big-number separator}
}
\value{
returns a character string that looks like x, but beautified
}
\description{
to be used inside the functions, just a convenient way to apply
  just-in-time formatting
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
