% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{tile_maker}
\alias{tile_maker}
\title{Tile maker
Tile maker}
\usage{
tile_maker(title = NULL, ...,
  css = "https://bootswatch.com/flatly/bootstrap.css")
}
\arguments{
\item{title}{Title.}

\item{...}{\code{div_maker} elements.}

\item{css}{A string indicating css url}
}
\description{
Tile maker
Tile maker
}
