% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{tile_matrix}
\alias{tile_matrix}
\title{tileMatrix}
\usage{
tile_matrix(values, titles = NULL, tar = 100, thre.H = 90, thre.L = 50,
  cols = 4, title, fileName, roundVal = 1, buttWidth = 100, margin = 3)
}
\arguments{
\item{values}{Vector containing values for each tile}

\item{titles}{Vector containing titles for each tile}

\item{tar}{Target value (What's the highest value to compare against). Defaults to 100}

\item{thre.H}{The limit between "high" and "medium" values IN PERCENT. Defaults to 90}

\item{thre.L}{The limit between "medium" and "low" values IN PERCENT. Defaults to 50}

\item{cols}{Number of columns that the matrix should tile around. Defaults to 4}

\item{title}{The title the matrix should have.}

\item{fileName}{The filename that will contain the html}

\item{roundVal}{Number of decimals that Value will be rounded to. Defaults to 1}

\item{buttWidth}{The width of each button element, in Number of pixels. Defaults to 100.}

\item{margin}{The amount of margin desired between buttons in pixels. Defaults to 3.}
}
\value{
Returns an HTML object containing the matrix of buttons
}
\description{
Create a matrix of buttons suitable for quick comparisons
}
\examples{
tile_matrix(c(3,4,5),c("Bob","Pedro","Ana"))
}
