% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilemaker.R
\name{TileMaker}
\alias{TileMaker}
\title{TileMaker}
\usage{
TileMaker(MainTitle = "", Divs, FileName, ShowDate = FALSE,
  localCSS = FALSE)
}
\arguments{
\item{MainTitle}{Optional title for the whole set of titles}

\item{Divs}{The Divs that you want inserted into the tile. If you have more than one div,
use `paste(Div1,Div2)`}

\item{FileName}{The filename the tile should spit out as, including the extension (which
should always be html)}

\item{ShowDate}{Optional boolean controlling whether the date should be included or suppressed}

\item{localCSS}{Optional boolean to specify whether the bootstrap css file should be served from
the internet, or if you have saved a local version. If you have saved a local version, make sure
to download the "fonts" folder too, otherwise glyphicons won't work.}
}
\value{
Use this function to output an html file containing all the divs.If you would like just
HTML code (suitable for inserting in a dashboard or another document, you can use the Divs.
}
\description{
Function 3 of 3, the last step. This function grabs the Divs created by `DivMaker` and
combines them into an html file.
}
\examples{
Div1 <- '<div class=\\"container\\"><h2>Quantativity factors</h2><a role=\\"button\\"  type=\\"button\\" class=\\"btn btn-warning btn-lg\\"><h1>
<span class=\\"glyphicon glyphicon-apple\\" aria-hidden=\\"true\\" style=\\"opacity:0.5\\"></span> 65<sup style= \\"font-size: 10px;
color:#EEEEEE\\">&#9650;1.6\%</sup></h1>Times apple eaten</a> <button type=\\"button\\" class=\\"btn btn-success btn-lg\\" title=\\" Apples are
high in good things. If you eat em, its nice. and further the cycle of good. Therefore, you should eat apples. \\" ><h1>93.7mg
<sup style= \\"font-size: 10px;color:#EEEEEE\\">&#9660;6.3\%</sup></h1>
Nutritional value</button></div>'
Div2 <- '<div class=\\"container\\"><h2>Implementation procedures</h2></div>'
## The two Divs above were built using this code... uncomment to see for yourself & experiment
# Button1 <- ButtonMaker(Color = 2,Value = 65,Former=64,Subtitle = "Times apple eaten",
#                        Link = "https://en.wikipedia.org/wiki/Apple",
#                        Icon="glyphicon glyphicon-apple",
#                        Target=100,ThresholdHigh=70,ThresholdLow=40)
# Button2 <- ButtonMaker(Color = 3,Value = 93.7,Former=100,Subtitle = "Nutritional value",
#                        Target=100,ThresholdHigh=70,ThresholdLow=40,Units="mg",
#                        Hover="Apples are high in good things. If you eat em, its nice.
#                        and further the cycle of good. Therefore, you should
#                        eat
#                        apples.")
# Button3 <- ButtonMaker(Color = 4,Value = 1,Subtitle = "Yumminess factor",
#                        Hover="While the yuminess is inferior to chocolate,
#' #                        it's still pretty friggin yummy.")
# Button4 <- ButtonMaker(Color = 5,Size=1,Value = 5,Subtitle = "Inconsistencies",
#                        Icon = "glyphicon glyphicon-pushpin")
#
# Div1 <- DivMaker(Title = "Quantativity factors",Buttons = paste(Button1,Button2))
# Div2 <- DivMaker(Title = "Implementation procedures",Buttons = paste(Button3,Button4))
TileMaker(MainTitle = "Hello",Divs = paste(Div1,Div2),FileName = "example.html")
browseURL("example.html")


}
\references{
Uses Twitter's awesome bootstrap V3
}
