% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilemaker.R
\name{tileMatrix}
\alias{tileMatrix}
\title{FUNCTION TITLE}
\usage{
tileMatrix(df, Tar = 100, Thre.H = 90, Thre.L = 50, cols = 4, Title,
  FileName, RoundVal = 1, ButtWidth = 100)
}
\arguments{
\item{df}{DESCRIPTION.}

\item{Tar}{DESCRIPTION.}

\item{Thre.H}{DESCRIPTION.}

\item{Thre.L}{DESCRIPTION.}

\item{cols}{DESCRIPTION.}

\item{Title}{DESCRIPTION.}

\item{FileName}{DESCRIPTION.}

\item{RoundVal}{DESCRIPTION.}

\item{ButtWidth}{DESCRIPTION.}
}
\value{
RETURN DESCRIPTION
}
\description{
FUNCTION DESCRIPTION
}
\examples{
# ADD EXAMPLES HERE
}
