% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{ico}
\alias{ico}
\title{Auxiliary function to generate icons}
\usage{
ico(x, chevron = FALSE)
}
\arguments{
\item{x}{Icon name. See http://getbootstrap.com/components/}
}
\description{
Auxiliary function to generate icons
}
