% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{multi_box}
\alias{multi_box}
\title{multi_box}
\usage{
multi_box(MB_values = NULL, mainTitle = NULL, size = "md",
  MB_icons = NULL, type = "info", link = NULL, MB_units = NULL,
  hover = NULL)
}
\arguments{
\item{MB_values}{vector of values to display, Default: NULL}

\item{mainTitle}{Top title, Default: NULL}

\item{size}{Optional size specified in the bootstrap css classes:
\"xs\",\"sm\",\"md\",\"lg\")}

\item{MB_icons}{vector of Icons to display, Default: NULL}

\item{type}{Optional bootstrap css element that governs the color. https://v4-alpha.getbootstrap.com/utilities/colors/
Choose from: "Muted", "Primary", "Success", "Info", "Warning", "Danger", Default: 'info'}

\item{link}{Optional hyperlink to redirect to after a user click, Default: NULL}

\item{MB_units}{vector of Units that explain the values, Default: NULL}

\item{hover}{Optional tooltip, or text that will show up when a user rests their
mouse over the tile, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a tile that contains more than one value, icon and units
}
\details{
DETAILS
}
\examples{
file_maker(multi_box(MB_values=c(3,45),mainTitle = "Important <br>button",
MB_icons=c("apple","calendar"), type="warning", MB_units=c("times","reports")))
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
