% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{multi_box}
\alias{multi_box}
\title{multi_box}
\usage{
multi_box(icons = NULL, txt = NULL, values = NULL, title = NULL,
  size = "md", type = "info", link = NULL, number_zoom = 150,
  hover = NULL, ...)
}
\arguments{
\item{icons}{vector of Icons to display, Default: NULL}

\item{txt}{Optional subtext that should appear under the value}

\item{values}{vector of values to display, Default: NULL}

\item{title}{Top title, Default: NULL}

\item{size}{Optional size specified in the bootstrap css classes:
"xs","sm","md","lg")}

\item{type}{Optional bootstrap css element that governs the color. https://v4-alpha.getbootstrap.com/utilities/colors/
Choose from: "Muted", "Primary", "Success", "Info", "Warning", "Danger", Default: 'info'}

\item{link}{Optional hyperlink to redirect to after a user click, Default: NULL}

\item{number_zoom}{Optional magnification \% for number vs normal text, Default: 150}

\item{hover}{Optional tooltip, or text that will show up when a user rests their
mouse over the tile, Default: NULL}

\item{...}{add any other html code here}
}
\value{
an HTML object
}
\description{
Create a tile that contains more than one value, icon and text
}
\details{
Allows for each button to contain several icon-number-text descriptions.
}
\examples{
library(dplyr)
multi_box(values=c(3,45),title = "Important <br>button",number_zoom=300,
icons=c("apple","calendar"), type="warning", txt=c("times","reports")) \%>\%
file_maker
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
