% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{solo_box}
\alias{solo_box}
\title{solo_box}
\usage{
solo_box(value = NULL, txt = NULL, former = NULL, size = "md",
  icon = NULL, color = "info", link = NULL, units = NULL,
  hover = NULL, textModifier = "h1", ...)
}
\arguments{
\item{value}{The numeric value you want to highlight (the main enchilada)}

\item{txt}{Optional subtext that should appear under the value}

\item{former}{The numeric old value to use for comparison to 'value'}

\item{size}{Optional size specified in the bootstrap css classes:
"xs","sm","md","lg")}

\item{icon}{Optional glyphicon that should be displayed from http://getbootstrap.com/components/ you need only supply
the name of thing you want, like "check"... not the full "gyphicon-check"}

\item{color}{Optional bootstrap css element that governs the color. https://v4-alpha.getbootstrap.com/utilities/colors/
Choose from: "Muted", "Primary", "Success", "Info", "Warning", "Danger"}

\item{link}{Optional hyperlink that should be followed on click}

\item{units}{Optional units that should be displayed after Value}

\item{hover}{Optional tooltip, or text that will show up when a user rests their mouse over the tile.}

\item{textModifier}{Optional css category of "large" text. In this case, the icon, value and unit.
In this case, title. Default=h1}

\item{...}{Optional additional html elements}
}
\description{
This function crafts the actual tile per se, including the specific
aesthetic traits for each tile. This is the simple version where you explicitly state the color.
}
\examples{
b1 <- solo_box(color= "warning", value = 3.57, txt = "B")
b2 <- solo_box(color= "danger", value = 13.7, txt = "Nutritional value")
b3 <- solo_box(color= "success", value = 1, txt = "Yumminess factor")
b4 <- solo_box(value = 3.57, former = 3, txt = "Times apple eaten", icon = "apple")
finisher(title = "straight buttons", divs = b1)
finisher(
  title = "with divs",
  divs = div_maker(
    subtitle = "boom",
    textModifier = "h1",
    div_maker(subtitle = "Boom", textModifier = "hi", b1, b2, b3)
  )
)

## Or taking advantage of the ability to change the textModifier:
finisher(
  title = "h4 modifier",
  divs = solo_box(
    value = 3, txt = "uh huh",
    former = 2, textModifier = "h4"
  )
)
}
