% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{solo_box}
\alias{solo_box}
\title{solo_box}
\usage{
solo_box(value = NULL, subtitle = NULL, former = NULL, size = "md",
  icon = NULL, type = "info", link = NULL, units = NULL, hover = NULL)
}
\arguments{
\item{value}{The numeric value you want to highlight (the main enchilada)}

\item{subtitle}{Optional subtext that should appear under the value}

\item{former}{The numeric old value to use for comparison to 'value'}

\item{size}{Optional size specified in the bootstrap css classes:
\"xs\",\"sm\",\"md\",\"lg\")}

\item{icon}{Optional glyphicon that should be displayed from http://getbootstrap.com/components/}

\item{type}{Optional bootstrap css element that governs the color. https://v4-alpha.getbootstrap.com/utilities/colors/
Choose from: "Muted", "Primary", "Success", "Info", "Warning", "Danger"}

\item{link}{Optional hyperlink that should be followed on click}

\item{units}{Optional units that should be displayed after Value}

\item{hover}{Optional tooltip, or text that will show up when a user rests their mouse over the tile.}
}
\description{
This function crafts the actual tile per se, including the specific
aesthetic traits for each tile. This is the simple version where you explicitly state the color.
}
\examples{
tile1 <- solo_box(type="warning",value = 3.57,subtitle = "B")
tile2 <- solo_box(type="danger",value = 13.7,subtitle = "Nutritional value")
tile3 <- solo_box(type="success",value = 1,subtitle = "Yumminess factor")
tile4 <- solo_box(value = 3.57, former=3,subtitle = "Times apple eaten", icon = "apple")
file_maker(tile1,tile2,tile3,tile4)
file_maker(div_maker(tile1,tile2,tile3),div_maker(tile4))
}
