% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{solo_gradient_box}
\alias{solo_gradient_box}
\title{solo_gradient_box}
\usage{
solo_gradient_box(value = NULL, subtitle = NULL, former = NULL,
  size = "md", icon = NULL, target = 100, thresholdHigh = 90,
  thresholdLow = 50, link = NULL, units = NULL, hover = NULL,
  hide_value = FALSE, ...)
}
\arguments{
\item{value}{The numeric value you want to highlight (the main enchilada)}

\item{subtitle}{Optional subtext that should appear under the value}

\item{former}{The last value that should be used for comparison purposes}

\item{size}{Optional size specified in the bootstrap css classes:
"xs","sm","md","lg")}

\item{icon}{Optional glyphicon that should be displayed from http://getbootstrap.com/components/}

\item{target}{Optional target that the value should be compared against. Use with ThresholdHigh and THresholdLow}

\item{thresholdHigh}{Optional edge between \"green\" and \"orange\" from 0-100. Use w/ Target and ThresholdLow. This value represents the RATIO
of the VALUE to the TARGET that, if above the ThresholdHigh will show as green, and if not, as orange}

\item{thresholdLow}{Optional border between \"orange\" and \"red\" from 0-100. Use w/ Target and ThresholdLow. This value represents the RATIO
of the VALUE to the TARGET that, if above the ThresholdHigh will show as orange, and if not, as red}

\item{link}{Optional hyperlink that should be followed on click}

\item{units}{Optional units that should be displayed after Value}

\item{hover}{Optional tooltip, or text that will show up when a user rests their mouse over the tile.}

\item{hide_value}{Optionally and paradoxically hide value. Normally FALSE, change this value to TRUE in order to suppress
the large number, but still take advantage of the conditional formatting.}

\item{...}{Optional additional html elements}
}
\description{
This function crafts a tile like solo_box, but this one changes color according to value
}
\examples{
# ADD EXAMPLES HERE
g1 <- solo_gradient_box(value = 40)
g2 <- solo_gradient_box(value = 40,target = 50,thresholdHigh = 80, thresholdLow=60)
g3 <- solo_gradient_box(value = 20,subtitle="Test1",target = 50,thresholdHigh = 80, thresholdLow=60,hide_value=T)
g4 <- solo_gradient_box(value = 35,subtitle="Test2",target = 50,thresholdHigh = 80, thresholdLow=60,hide_value=T)
file_maker(div_maker(g1,g2,g3,g4))
}
