% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_maker.R
\name{tile_matrix}
\alias{tile_matrix}
\title{tileMatrix}
\usage{
tile_matrix(values, titles, former = NULL, tar = 100, thre.H = 90,
  thre.L = 50, cols = 4, mainTitle = NULL, roundVal = 1)
}
\arguments{
\item{values}{Vector containing values for each tile}

\item{titles}{Vector containing titles for each tile}

\item{former}{vector containing former values (to show change from last)}

\item{tar}{Target value (What's the highest value to compare against). Defaults to 100}

\item{thre.H}{The limit between "high" and "medium" values IN PERCENT. Defaults to 90}

\item{thre.L}{The limit between "medium" and "low" values IN PERCENT. Defaults to 50}

\item{cols}{Number of columns that the matrix should tile around. Defaults to 4}

\item{mainTitle}{The title the matrix should have.}

\item{roundVal}{Number of decimals that Value will be rounded to. Defaults to 1}
}
\value{
Returns a list object containing the matrix of buttons
}
\description{
Create a matrix of buttons suitable for quick comparisons
}
\examples{
file_maker(tile_matrix(values=c(3,4,5),titles=c("Bob","Pedro","Ana")))
}
